--- camunda-console-client-java-core/src/main/java/io/camunda/console/client/exception/CamundaConsoleClientException.java
@@ -1,8 +1,0 @@
-package io.camunda.console.client.exception;
-
-public class CamundaConsoleClientException extends RuntimeException {
-
-  public CamundaConsoleClientException(Throwable cause) {
-    super(cause);
-  }
-}
