--- camunda-console-client-java-core/src/main/java/io/camunda/console/client/impl/AbstractCamundaConsoleClient.java
@@ -1,19 +1,0 @@
-package io.camunda.console.client.impl;
-
-import io.camunda.console.client.api.DefaultApi;
-
-public abstract class AbstractCamundaConsoleClient {
-  private final DefaultApi api;
-
-  public AbstractCamundaConsoleClient(DefaultApi api) {
-    this.api = api;
-  }
-
-  public AbstractCamundaConsoleClient(AbstractCamundaConsoleClient consoleClient) {
-    this.api = consoleClient.api;
-  }
-
-  protected DefaultApi getApi() {
-    return api;
-  }
-}
