--- camunda-console-client-java-core/src/test/resources/camunda-console-client-omit-defaults.properties
@@ -1,2 +1,0 @@
-camunda.console.client-id=clientId
-camunda.console.client-secret=clientSecret
