--- camunda-console-client-java-core/src/test/resources/camunda-console-client.properties
@@ -1,5 +1,0 @@
-camunda.console.client-id=clientId
-camunda.console.client-secret=clientSecret
-camunda.oauth-url=https://login.my.camunda.io/oauth/token
-camunda.console.base-url=https://api.my.camunda.io
-camunda.console.oauth-audience=api.my.camunda.io
