+++ camunda-management-api-client-java-core/src/main/java/io/camunda/management/api/client/exception/CamundaConsoleClientException.java
@@ -1,0 +1,8 @@
+package io.camunda.management.api.client.exception;
+
+public class CamundaConsoleClientException extends RuntimeException {
+
+  public CamundaConsoleClientException(Throwable cause) {
+    super(cause);
+  }
+}
