+++ camunda-management-api-client-java-core/src/main/java/io/camunda/management/api/client/impl/CamundaManagementApiClientImpl.java
@@ -1,0 +1,32 @@
+package io.camunda.management.api.client.impl;
+
+import io.camunda.management.api.client.api.CamundaManagementApiClient;
+import io.camunda.management.api.client.api.DefaultApi;
+
+public class CamundaManagementApiClientImpl extends AbstractCamundaManagementApiClient
+    implements CamundaManagementApiClient {
+
+  public CamundaManagementApiClientImpl(DefaultApi api) {
+    super(api);
+  }
+
+  @Override
+  public Clusters clusters() {
+    return new ClustersImpl(this);
+  }
+
+  @Override
+  public Cluster clusters(String id) {
+    return new ClusterImpl(this, id);
+  }
+
+  @Override
+  public Members members() {
+    return new MembersImpl(this);
+  }
+
+  @Override
+  public Member members(String email) {
+    return new MemberImpl(this, email);
+  }
+}
