+++ camunda-management-api-client-java-core/src/test/resources/camunda-console-client-omit-defaults.properties
@@ -1,0 +1,2 @@
+camunda.management.api.client-id=clientId
+camunda.management.api.client-secret=clientSecret
