+++ camunda-management-api-client-java-core/src/test/resources/camunda-console-client.properties
@@ -1,0 +1,5 @@
+camunda.management.api.client-id=clientId
+camunda.management.api.client-secret=clientSecret
+camunda.oauth-url=https://login.my.camunda.io/oauth/token
+camunda.management.api.base-url=https://api.my.camunda.io
+camunda.management.api.oauth-audience=api.my.camunda.io
