+++ pom.xml
@@ -7,12 +7,12 @@
     <version>1.4.4</version>
   </parent>
   <groupId>io.camunda</groupId>
-  <artifactId>camunda-console-client-java-parent</artifactId>
+  <artifactId>camunda-management-api-client-java-parent</artifactId>
   <version>1.2-SNAPSHOT</version>
   <packaging>pom</packaging>
   <modules>
-    <module>camunda-console-client-java-generated</module>
-    <module>camunda-console-client-java-core</module>
+    <module>camunda-management-api-client-java-generated</module>
+    <module>camunda-management-api-client-java-core</module>
   </modules>
 
   <properties>
@@ -68,7 +68,7 @@
       </dependency>
       <dependency>
         <groupId>io.camunda</groupId>
-        <artifactId>camunda-console-client-java-generated</artifactId>
+        <artifactId>camunda-management-api-client-java-generated</artifactId>
         <version>${project.version}</version>
       </dependency>
     </dependencies>
